+++ packages/binder/lib/src/binder_scope.dart
@@ -203,7 +203,7 @@ class BinderScopeState extends State<BinderScope>
   }
 
   @override
-  S read<T, S>(Watchable<T, S> ref) {
+  T read<T>(Watchable<T> ref) {
     return ref.read(fetch);
   }