+++ packages/binder/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0]
+### Added
+- `Consumer` widget used to watch one watchable.
+
+### Modified
+- `Watchable` has now only one generic type.
+
 ## [0.1.5] 
 ### Modified
 - Only calls `Aspect.shouldRebuild` when necessary.
@@ -31,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/letsar/binder/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/letsar/binder/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/letsar/binder/compare/releases/tag/v0.2.0
 [0.1.5]: https://github.com/letsar/binder/compare/releases/tag/v0.1.5
 [0.1.4]: https://github.com/letsar/binder/compare/releases/tag/v0.1.4
 [0.1.3]: https://github.com/letsar/binder/compare/releases/tag/v0.1.3